# Use carrierwave-vips for faster image transcoding