--- a/Aptfile
+++ b/Aptfile
@@ -1,0 +1,4 @@
+# Install dependencies required by heroku-buildpack-vips
+libglib2.0-0
+libglib2.0-dev
+libpoppler-glib8