--- a/app/uploaders/media_file_uploader.rb
+++ b/app/uploaders/media_file_uploader.rb
@@ -3,7 +3,7 @@
 # Such as MP3 tracks or PDF documents
 class MediaFileUploader < ApplicationUploader
 
-  include CarrierWave::MiniMagick
+  include CarrierWave::Vips
 
   VERSIONS = {
     huge: 2880,