… Gemfile.lock
… test/fixtures/files/image.jpg