Not ready for merge yet. Needs some profiling and acceptance testing. Also likely need to make libvips available on Travis + Heroku.

Find below some initial results from uploading the test files on my local machine. TLDR; one is randomly a bit faster than the other so no distinct gains.

Will run the tests on staging on heroku when I have time to see if there are any variables in that environment that would better distinguish one from the other.

## Camel

### Vips

Completed 200 OK in 5208ms (Views: 0.3ms | ActiveRecord: 40.5ms | Allocations: 57706)
Completed 200 OK in 5069ms (Views: 0.2ms | ActiveRecord: 11.9ms | Allocations: 55811)
Completed 200 OK in 5116ms (Views: 0.2ms | ActiveRecord: 11.1ms | Allocations: 55832)

### MiniMagick

Completed 200 OK in 4665ms (Views: 0.2ms | ActiveRecord: 12.8ms | Allocations: 63043)
Completed 200 OK in 4717ms (Views: 0.2ms | ActiveRecord: 11.6ms | Allocations: 63080)
Completed 200 OK in 4492ms (Views: 0.1ms | ActiveRecord: 11.6ms | Allocations: 59342)

## Flowers

### Vips

Completed 200 OK in 4016ms (Views: 0.2ms | ActiveRecord: 11.6ms | Allocations: 59561)
Completed 200 OK in 3877ms (Views: 0.2ms | ActiveRecord: 12.4ms | Allocations: 55810)
Completed 200 OK in 3926ms (Views: 0.2ms | ActiveRecord: 11.9ms | Allocations: 55803)

### MiniMagick

Completed 200 OK in 4401ms (Views: 0.2ms | ActiveRecord: 11.4ms | Allocations: 59311)
Completed 200 OK in 4126ms (Views: 0.2ms | ActiveRecord: 11.2ms | Allocations: 59338)
Completed 200 OK in 3956ms (Views: 0.2ms | ActiveRecord: 11.7ms | Allocations: 59286)

## Ice Cream

### Vips

Completed 200 OK in 3353ms (Views: 0.2ms | ActiveRecord: 11.3ms | Allocations: 55798)
Completed 200 OK in 3299ms (Views: 0.2ms | ActiveRecord: 11.2ms | Allocations: 55800)
Completed 200 OK in 3350ms (Views: 0.2ms | ActiveRecord: 11.8ms | Allocations: 55794)

### MiniMagick

Completed 200 OK in 4081ms (Views: 0.2ms | ActiveRecord: 11.9ms | Allocations: 59315)
Completed 200 OK in 4311ms (Views: 0.2ms | ActiveRecord: 11.7ms | Allocations: 59293)
Completed 200 OK in 4004ms (Views: 0.2ms | ActiveRecord: 11.9ms | Allocations: 59288)

## Pallazo

### Vips

Completed 200 OK in 4774ms (Views: 0.2ms | ActiveRecord: 11.0ms | Allocations: 55803)
Completed 200 OK in 4565ms (Views: 0.1ms | ActiveRecord: 11.2ms | Allocations: 55798)
Completed 200 OK in 4591ms (Views: 0.2ms | ActiveRecord: 11.4ms | Allocations: 55817)

### MiniMagick

Completed 200 OK in 4606ms (Views: 0.2ms | ActiveRecord: 11.4ms | Allocations: 59302)
Completed 200 OK in 4458ms (Views: 0.2ms | ActiveRecord: 11.1ms | Allocations: 59303)
Completed 200 OK in 4335ms (Views: 0.2ms | ActiveRecord: 12.8ms | Allocations: 59281)