@dbackeus 2019-12-06
Closing this since huge performance gains appear unlikely and I can't even find a Heroku buildpack supporting `ruby-vips` that also supports webp transformations.

On a final note I did see a scenario where Vips was consistently 2x faster than ImageMagick in my local environment. But that only happened if turning off WebP conversions. So it would appear that Vips handles jpg much better than Vips but might be worse at WebP.